changed -t int to float

+++ patristic_distances_filtered.py
@@ -12,7 +12,7 @@ def parseArgs():
     parser = argparse.ArgumentParser(description=__doc__, formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('-i', '--input', required=False, type=argparse.FileType('r'), default=stdin, help="Input FASTA")
     parser.add_argument('-o', '--output', required=False, type=argparse.FileType('w'), default=stdout, help="Output, don't include file type")
-    parser.add_argument('-t', '--threshold', required=False, type=int, default=1, help="Threshold")
+    parser.add_argument('-t', '--threshold', required=False, type=float, default=1, help="Threshold")
     args = parser.parse_args()
     return args.input, args.output, args.threshold
